+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - feature: add info whether the user has role user and admin
 - feature: rename customer to staff
--
+
 ## 1.0.0
 
 - First Release