+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feature: add info whether the user has role user and admin
 - feature: rename customer to staff
+- refactor: rename UserFixtures to AppFixtures
 
 ## 1.0.0