+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feature: add info whether the user has role user and admin
 - feature: rename customer to staff
 - refactor: rename UserFixtures to AppFixtures
+- build: add linters
 
 ## 1.0.0