+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feature: rename customer to staff
 - refactor: rename UserFixtures to AppFixtures
 - build: add linters
+- feature: add trait for tests
 
 ## 1.0.0