+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - refactor: rename UserFixtures to AppFixtures
 - build: add linters
 - feature: add trait for tests
+- feature: JWT Refresh Token
 
 ## 1.0.0